requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
scipy>=1.12
matplotlib>=3.8
pyyaml>=6.0
joblib>=1.4
pyarrow>=15.0
tqdm>=4.66
rich>=13.7
